move test file and examples file out of subdirectory
