+++ examples.py
